pipeline/requirements.txt
requests
PyYAML
psycopg2-binary
pandas
